% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-cis.R
\name{compute_oe_ci}
\alias{compute_oe_ci}
\title{Compute two-sided confidence intervals for Poisson parameters.}
\usage{
compute_oe_ci(o, e, conf.level = 0.9, midp = TRUE, value_if_e0 = NA_real_,
  warn_if_e0 = TRUE)
}
\arguments{
\item{o}{the observed number of events (a vector of non-negative integers)}

\item{e}{the expected number of events (a vector of non-negative numbers)}

\item{conf.level}{the confidence level (a vector of numbers between 0 and 1,
with default value 0.90)}

\item{midp}{a vector of logicals.  If \code{TRUE} (the default), then compute
mid-p confidence intervals instead of exact confidence intervals.}

\item{value_if_e0}{This parameter determines the confidence interval for the
case \code{e = 0}.  If \code{value_if_e0} consists of a single
numerical value (default: \code{NA_real_}), both boundaries of the
confidence interval are set to this value.  If \code{value_if_e0} is a
vector of two values, these values are taken as lower and upper boundary
of the confidence interval.  If \code{NULL}, the interval is
[0, \eqn{\infty}].}

\item{warn_if_e0}{logical of length one.  If \code{TRUE} (the default), then
print a warning if \code{e = 0} occurs in the input.}
}
\value{
A \code{data.frame} with two columns, the \code{lower} and the
\code{upper} bound of the confidence interval.
}
\description{
Compute two-sided confidence intervals for the parameter \eqn{\lambda}, under
the assumption that the underlying distribution of \eqn{o} is Poisson with
parameter \eqn{e\cdot\lambda}{e·\lambda}.
}
\details{
The confidence interval at confidence level \code{conf.level} is computed by
inverting two one-sided tests, each with significance level
(1 - \code{conf.level})/2 (see \code{\link{compute_oe_pvalue}} for details).

The case \eqn{e=0} deserves particular attention.  In this degenerated case,
the parameter \eqn{\lambda} and the confidence interval become meaningless.
Therefore, the default behaviour is to assign a value of \code{NA_real_} to
both boundaries of the confidence interval and to print a warning.  The
arguments \code{value_if_e0} and \code{warn_if_e0} can be used to change this
behaviour.

The arguments \code{o}, \code{e}, \code{conf.level} and \code{midp} may have
arbitrary lengths.  If the lengths differ, arguments are recycled according
to the usual \R rules (warnings may appear if lenghts are not multiples of
each other).  If any of these arguments has length zero, the result has
zero rows.  If any of these arguments is \code{NA}, the result will also be
\code{NA}.
}
\examples{
compute_oe_ci(o = 2, e = 1.8)

}
\seealso{
Other confidence interval functions: \code{\link{compute_rate_ci}}

Other functions for o/e indicators: \code{\link{compute_oe_pvalue}}
}
