% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-pvalues.R
\name{compute_oe_pvalue}
\alias{compute_oe_pvalue}
\title{Compute the p-values of exact one-sided Poisson tests.}
\usage{
compute_oe_pvalue(o, e, t_smr, alternative = "greater", midp = TRUE,
  value_if_e0 = NA_real_, warn_if_e0 = TRUE)
}
\arguments{
\item{o}{the observed number of events (a vector of non-negative integers)}

\item{e}{the expected number of events (a vector of non-negative numbers)}

\item{t_smr}{the threshold value of the null hypothesis (a vector of
non-negative numbers)}

\item{alternative}{direction of the alternative (a character vector with
elements \code{"greater"} for
\eqn{H_1: \lambda > t_{\mbox{SMR}}}{H_1: \lambda > t_smr} or
\code{"less"} for
\eqn{H_1: \lambda < t_{\mbox{SMR}}}{H_1: \lambda < t_smr}).
The default value is \code{"greater"}.}

\item{midp}{a vector of logicals.  If \code{TRUE} (the default), then compute
mid-p values instead of exact p-values}

\item{value_if_e0}{The (mid)pvalue for the case \code{e = 0} (a numeric of
length one or \code{NULL}).  The default  value is \code{NA_real_}.
If equal to \code{NULL}, then the p-value is equal to 1 and the mid-p-value
is equal to 0.5 when \eqn{o = e = 0}; otherwise, when \eqn{o > e = 0}, the
(mid)p-value is 0.}

\item{warn_if_e0}{logical of length one.  If \code{TRUE} (the default), then
print a warning if \code{e = 0} occurs in the input.}
}
\value{
A vector of (mid)p-values.
}
\description{
Compute the p-values of exact one-sided Poisson tests.  Suppose that the
underlying distribution of \eqn{o} is Poisson with parameter
\eqn{e\cdot\lambda}{e·\lambda}.
\ifelse{latex}{If \code{alternative = "greater"}, then null hypothesis and
 alternative hypothesis are given as
 \deqn{H_0: \lambda \le t_{\mbox{SMR}}
   \quad\mbox{and}\quad H_1: \lambda > t_{\mbox{SMR}}.}
 Conversely, if \code{alternative = "less"}, then null hypothesis and
 alternative hypothesis are given as
 \deqn{H_0: \lambda \ge t_{\mbox{SMR}}
  \quad\mbox{and}\quad H_1: \lambda < t_{\mbox{SMR}}.}
}{If \code{alternative = "greater"}, then the null hypothesis is
 \deqn{H_0: \lambda \le t_smr}
 with alternative hypothesis
 \eqn{H_1: \lambda > t_smr.}
 Conversely, if \code{alternative = "less"}, then the null hypothesis is
 \deqn{H_0: \lambda \ge t_smr}
 with alternative hypothesis
 \eqn{H_1: \lambda < t_smr.}
}
}
\details{
When \code{alternative = "greater"}, the p-value is defined as
the probability \eqn{\mbox{Pr}(X\ge o)}{Pr(X \ge o)}, where \eqn{X} is a
Poisson random variable with expected value \eqn{e\cdot\lambda}{e·\lambda}.
The corresponding mid-p-value is defined as
\deqn{\mbox{Pr}(X > o) + \frac12\mbox{Pr}(X = o).}{Pr(X > o) + ½ Pr(X = o).}

The case \eqn{e=0} deserves particular attention.  In this case, \eqn{o}
should also be zero; when \eqn{e=0} and \eqn{o>0}, then the null hypothesis
can be immediately rejected, which corresponds to a (mid)p-value of zero.
When \eqn{o=e=0}, then, formally, the p-value is 1, and the mid-p-value is
0.5.  However, when \eqn{e=0}, this usually means that there is no data
available to distinguish between the two hypotheses, and so the (mid)p-value
is meaningless.  Therefore, the default behaviour is to assign a value of
\code{NA_real_} to this case and to print a warning.  The arguments
\code{value_if_e0} and \code{warn_if_e0} can be used to change this
behaviour.

The arguments \code{o}, \code{e}, \code{t_smr} and \code{alternative} may have
arbitrary lengths.  If the lengths differ, arguments are recycled according
to the usual \R rules (warnings may appear if lenghts are not multiples of
each other).  If any of these arguments has length zero, the result also has
length zero.  If any of these arguments is \code{NA}, the result will also be
\code{NA}.
}
\examples{
compute_oe_pvalue(o = 2, e = 1.8, t_smr = 2.18)

}
\seealso{
Other p-value functions: \code{\link{compute_rate_pvalue}}

Other functions for o/e indicators: \code{\link{compute_oe_ci}}
}
