% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-pvalues.R
\name{compute_rate_pvalue}
\alias{compute_rate_pvalue}
\title{Compute (mid)p-values of exact one-sided binomial tests.}
\usage{
compute_rate_pvalue(o, n, t, alternative = "greater", midp = TRUE,
  value_if_n0 = NA_real_, warn_if_n0 = TRUE)
}
\arguments{
\item{o}{the observed number of events (a vector of non-negative integers)}

\item{n}{total number of cases (a vector of non-negative integers that
satisfies \code{o <= n})}

\item{t}{the threshold value of the null hypothesis (a vector of numbers
between 0 and 1)}

\item{alternative}{direction of the alternative (a character vector with
elements \code{"greater"} for \eqn{H_1: p>t} or
\code{"less"} for \eqn{H_1: p<t}).  The default value is \code{"greater"}.}

\item{midp}{a vector of logicals.  If \code{TRUE} (the default), then compute
mid-p values instead of exact p-values.}

\item{value_if_n0}{The (mid-)p-value for the case \code{o = n = 0} (a numeric
of length one or \code{NULL}).  The default  value is \code{NA_real_}.
If equal to \code{NULL}, then the p-value is equal to 1 and the mid-p-value
is equal to 0.5.}

\item{warn_if_n0}{logical of length one.  If \code{TRUE} (the default), then
print a warning if the combination \code{o = n = 0} occurs in the input.}
}
\value{
A vector of (mid)p-values.
}
\description{
Compute (mid)p-values of exact one-sided binomial tests.
\ifelse{latex}{If \code{alternative = "greater"}, then null hypothesis and
 alternative hypothesis are given as
 \deqn{H_0: p \le t \quad\mbox{and}\quad H_1: p > t.}
 Conversely, if \code{alternative = "less"}, then null hypothesis and the
 alternative hypothesis are given as
 \deqn{H_0: p \ge t \quad\mbox{and}\quad H_1: p < t.}
}{
 If \code{alternative = "greater"}, then the null hypothesis is
 \deqn{H_0: p \le t}
 with alternative hypothesis
 \eqn{H_1: p > t.}
 Conversely, if \code{alternative = "less"}, then the null hypothesis is
 \deqn{H_0: p \ge t}
 with alternative hypothesis
 \eqn{H_1: p < t.}
}
}
\details{
When \code{alternative = "greater"}, the p-value is defined as
the probability \eqn{\mbox{Pr}(X\ge o)}{Pr(X \ge o)}, where \eqn{X} is a
binomial random variable with \eqn{n} trials and success probability \eqn{t}.
The corresponding mid-p-value is defined as
\deqn{\mbox{Pr}(X > o) + \frac12\mbox{Pr}(X = o).}{Pr(X > o) + ½ Pr(X = o).}

The case \eqn{o=n=0} deserves particular attention.  Formally, the
p-value is 1, and the mid-p-value is 0.5 in this case.  However, when
\eqn{n=0}, then there is no data available to distinguish between the two
hypotheses, and so the (mid)p-value is meaningless.  Therefore, the default
behaviour is to assign a value of \code{NA_real_} to this case and to print a
warning.  The arguments \code{value_if_n0} and \code{warn_if_n0} can be
used to change this behaviour.

The arguments \code{o}, \code{n}, \code{t} and \code{alternative} may have
arbitrary lengths.  If the lengths differ, arguments are recycled according
to the usual \R rules (warnings may appear if lenghts are not multiples of
each other).  If any of these argument has length zero, the result also has
length zero.  If any of these argument is \code{NA}, the result will also be
\code{NA}.
}
\examples{
compute_rate_pvalue(o = 1, n = 2, t = 0.95, alternative = "less")

}
\seealso{
Other p-value functions: \code{\link{compute_oe_pvalue}}

Other functions for rate indicators: \code{\link{compute_rate_ci}}
}
